pyproject.toml
[project]
name = "etic"
version = "0.1.13"
description = "Easily download bibtex formatted citations from DOI identifier."
readme = "README.md"
authors = [
    { name = "dasdias", email = "das.dias6@gmail.com" }
]
license = "MIT"

requires-python = ">=3.10"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.13",
    "asyncio>=3.4.3",
    "bibtexparser>=1.4.3",
    "docopt>=0.6.2",
    "pathlib>=1.0.1",
    "pyyaml>=6.0.2",
]

[project.scripts]
etic = "etic:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# The path is relative to the project root
include = [
    "/src/etic/config.yml"
]

